Added mech and camera functionality. UI base class, drawing the mouse cursor. Game object and message system updates. Few other (minor) updates.
git-svn-id: file:///home/inzane/.dbot-svn-fix/trunk@289 d4333498-d51f-0410-a6fe-c081ca229757

+++ game/game/include/heightMap.h
@@ -21,7 +21,7 @@ public:
 	float height(float aX, float aY);
 private:
 	int mHMapSize, mVMapSize;
-	int **mHeightMap;
+	float **mHeightMap;
 	float mHeight;
 	int mMaxHeight;
 	int mColorNum;
